Split plugins
